+++ crates/azure_types/src/lib.rs
@@ -17,6 +17,9 @@ pub mod cost_management;
 mod duration;
 mod eligible_child_resources;
 mod groups;
+mod key_vault;
+mod key_vault_id;
+mod key_vault_name;
 mod management_group_ancestors_chain;
 mod management_groups;
 mod metrics;
@@ -101,9 +104,11 @@ mod virtual_network_name;
 mod virtual_network_peering_id;
 mod virtual_network_peering_name;
 mod virtual_network_properties;
+mod key_vault_properties;
 
 pub mod prelude {
     pub use crate::accounts::*;
+    pub use crate::key_vault_properties::*;
     pub use crate::address_prefix::*;
     pub use crate::address_prefixes::*;
     pub use crate::all_or::*;
@@ -120,6 +125,9 @@ pub mod prelude {
     pub use crate::duration::*;
     pub use crate::eligible_child_resources::*;
     pub use crate::groups::*;
+    pub use crate::key_vault::*;
+    pub use crate::key_vault_id::*;
+    pub use crate::key_vault_name::*;
     pub use crate::management_group_ancestors_chain::*;
     pub use crate::management_groups::*;
     pub use crate::metrics::*;