+++ crates/entrypoint/src/interactive/copy_azurerm_backend_menu.rs
@@ -30,11 +30,11 @@ pub async fn copy_azurerm_backend_menu() -> Result<()> {
             .into_iter()
             .map(|sa| {
                 let sub_name = subscriptions
-                    .get(&sa.subscription_id)
+                    .get(&sa.id.resource_group_id.subscription_id)
                     .map(|sub| sub.name.to_owned())
                     .unwrap_or_else(|| SubscriptionName::try_new("Unknown Subscription").unwrap());
-                let key = format!("{:32}\t{:64}\t{}", sub_name, sa.resource_group, sa.name);
-                let key_short = format!("{} {} {}", sub_name, sa.resource_group, sa.name);
+                let key = format!("{:32}\t{:64}\t{}", sub_name, sa.id.resource_group_id.resource_group_name, sa.name);
+                let key_short = format!("{} {} {}", sub_name, sa.id.resource_group_id.resource_group_name, sa.name);
                 Choice {
                     key,
                     value: (sa, key_short, sub_name),
@@ -51,7 +51,7 @@ pub async fn copy_azurerm_backend_menu() -> Result<()> {
     cmd.args(["storage", "container", "list", "--account-name"]);
     cmd.arg(&*chosen_storage_account.0.name);
     cmd.arg("--subscription");
-    cmd.arg(chosen_storage_account.0.subscription_id.short_form());
+    cmd.arg(chosen_storage_account.0.id.resource_group_id.subscription_id.short_form());
     cmd.args(["--query", "[].name", "--output", "json"]);
     let blob_container_names = cmd.run::<Vec<String>>().await?;
 
@@ -85,10 +85,10 @@ pub async fn copy_azurerm_backend_menu() -> Result<()> {
     container_name       = "{}"
     subscription_id      = "{}" # {}
     "#,
-        chosen_storage_account.0.resource_group,
+        chosen_storage_account.0.id.resource_group_id.resource_group_name,
         chosen_storage_account.0.name,
         chosen_blob_container,
-        chosen_storage_account.0.subscription_id.short_form(),
+        chosen_storage_account.0.id.resource_group_id.subscription_id.short_form(),
         chosen_storage_account.2
     );